requirements.txt
Django==3.0.1
python_bcrypt==0.3.2
